pyproject.toml
[tool.poetry]
name = "compliance-screenshot-archiver"
version = "0.1.0"
description = "Scheduled and on-demand webpage captures with verifiable artifacts."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
pydantic = "^2.8.2"
uvicorn = "^0.30.0"
python-dotenv = "^1.0.1"
boto3 = "^1.34.0"
mangum = "^0.17.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
ruff = "^0.5.0"
mypy = "^1.10.0"
pytest = "^8.2.0"
pytest-asyncio = "^0.23.8"
moto = "^5.0.10"
types-requests = "*"
bandit = "^1.7.9"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
lint.select = ["E","F","I","UP","B","S","W","PL","N"]

[tool.mypy]
python_version = "3.12"
strict = true
exclude = ["tests/"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
